move tests require -> import
