+++ themes-default/slim/test/specs/add-shows.spec.js
@@ -2,11 +2,8 @@ import test from 'ava';
 import Vuex from 'vuex';
 import VueRouter from 'vue-router';
 import { createLocalVue, mount } from '@vue/test-utils';
+import { AppLink, AddShows } from '../../static/js/templates';
 import fixtures from '../__fixtures__/common';
-
-// Needs to be required otherwise nyc won't see it
-const AppLink = require('../../static/js/templates/app-link.vue');
-const AddShows = require('../../static/js/templates/add-shows.vue');
 
 test.beforeEach(t => {
     t.context.localVue = createLocalVue();
