+++ themes-default/slim/test/specs/anidb-release-group-ui.spec.js
@@ -2,10 +2,8 @@ import test from 'ava';
 import Vuex from 'vuex';
 import VueRouter from 'vue-router';
 import { createLocalVue, mount } from '@vue/test-utils';
+import { AnidbReleaseGroupUi } from '../../static/js/templates';
 import fixtures from '../__fixtures__/common';
-
-// Needs to be required otherwise nyc won't see it
-const AnidbReleaseGroupUi = require('../../static/js/templates/anidb-release-group-ui.vue');
 
 test.beforeEach(t => {
     t.context.localVue = createLocalVue();
