+++ themes-default/slim/test/specs/http/404.spec.js
@@ -1,8 +1,6 @@
 import test from 'ava';
 import { mount } from '@vue/test-utils';
-
-// Needs to be required otherwise nyc won't see it
-const NotFound = require('../../../static/js/templates/http/404.vue');
+import { NotFound } from '../../../static/js/templates';
 
 test('renders not-found page', t => {
     const wrapper = mount(NotFound);
