+++ themes-default/slim/test/specs/scroll-buttons.spec.js
@@ -1,8 +1,6 @@
 import test from 'ava';
 import { createLocalVue, mount } from '@vue/test-utils';
-
-// Needs to be required otherwise nyc won't see it
-const ScrollButtons = require('../../static/js/templates/scroll-buttons.vue');
+import { ScrollButtons } from '../../static/js/templates';
 
 test.beforeEach(t => {
     t.context.localVue = createLocalVue();
